+++ packages/components/modules/messages/common/graphql/fragments/ProfileSummary.ts
@@ -4,15 +4,16 @@ export const ProfileSummaryFragment = graphql`
   fragment ProfileSummaryFragment on ChatRoom
   @argumentDefinitions(
     avatarSize: { type: "Int", defaultValue: 100 }
+    cursor: { type: "String" }
     count: { type: "Int", defaultValue: 5 }
   ) {
     id
-    participants(first: $count) {
+    participants(first: $count, after: $cursor)
+      @connection(key: "ProfileSummaryFragment_participants") {
       edges {
         node {
           profile {
             id
-            pk
             name
             image(width: $avatarSize, height: $avatarSize) {
               url