+++ packages/components/modules/messages/web/MessagesList/MessagesGroup/UserMessage/MessageItem/__storybook__/mockResolvers.ts
@@ -8,7 +8,6 @@ export const mockResolvers = {
     extraData: null,
     inReplyTo: { id: 'msg-0' },
     isRead: true,
-    pk: 12345,
     profile: { id: 'profile-123' },
     verb: 'SENT_MESSAGE',
   }),