+++ packages/design-system/components/native/appbars/AppBar/types.ts
@@ -3,9 +3,11 @@ import { FC } from 'react'
 import { SvgIconProps } from '../../icons'
 
 export type AppBarProps = {
-  title: string
+  title?: string
+  titleComponent?: React.ReactNode
   onBack?: () => void
   onClose?: () => void
   BackIcon?: FC<SvgIconProps>
   CloseIcon?: FC<SvgIconProps>
+  closeComponent?: React.ReactNode
 }