+++ packages/eslint-plugin/.eslintrc.js
@@ -1,1 +1,21 @@
-module.exports = require('@baseapp-frontend/config/.eslintrc.js')
+module.exports = {
+  env: {
+    node: true,
+    es2021: true,
+  },
+  parserOptions: {
+    ecmaVersion: "latest",
+    sourceType: "module",
+  },
+  extends: ["eslint:recommended"],
+  rules: {
+    "no-unused-vars": [
+      "error",
+      {
+        argsIgnorePattern: "^_",
+        varsIgnorePattern: "^_",
+        caughtErrorsIgnorePattern: "^_",
+      },
+    ],
+  },
+};