+++ packages/eslint-plugin/index.ts
@@ -1,3 +1,0 @@
-import rules from './rules'
-
-export { rules }